+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.35.0] - 2025-05-??
+
+### Removed
+
+- Removing `firmware.structs` Ragger-specific device definitions
+
+### Changed
+
+- Using `ledgered` Devices instead of local `Firmware` / `FIRMWARES`.
+  Basically `ledgered.device.DeviceType` replaces the `ragger.firmware.Firmware` enum, and
+  `ledgered.device.Devices` replaces `FIRMWARES`.
+
 ## [1.34.0] - 2025-04-28
 
 ### Changed